+++ firmware/Pic32/RFIDler.X/include/q5.h
@@ -15,7 +15,7 @@
  * o RFIDler-LF Nekkid                                                     *
  *                                                                         *
  *                                                                         *
- * RFIDler is (C) 2013-2014 Aperture Labs Ltd.                             *
+ * RFIDler is (C) 2013-2015 Aperture Labs Ltd.                             *
  *                                                                         *
  * This program is free software; you may redistribute and/or modify it    *
  * under the terms of the GNU General Public License as published by the   *
@@ -208,4 +208,4 @@ BOOL q5_login(BYTE *response, BYTE *pass);
 BOOL q5_rwd_test(BYTE *pattern);
 BOOL q5_config_block_show(BYTE *config, BYTE *password);
 BOOL q5_emulate_config_block(BYTE *config, BYTE target_tagtype);
-BOOL q5_decode_pwm(unsigned long pulses[], unsigned long gaps[], unsigned int count);
+BOOL q5_decode_pwm(unsigned int pulses[], unsigned int gaps[], unsigned int count);
