+++ firmware/Pic32/RFIDler.X/include/sniff.h
@@ -130,5 +130,5 @@
 // Author: Adam Laurie <adam@aperturelabs.com>
 
 void sniff_pwm(unsigned int mingap, unsigned int minpulse, unsigned int mesg_gap, BOOL nfcmode);
-void decode_pwm(unsigned long pulses[], unsigned long gaps[], unsigned int count);
-BYTE generic_decode_pwm(BYTE *result, unsigned long pulses[], unsigned int minpulse, unsigned int maxpulse, unsigned long gaps[], unsigned int mingap, unsigned int maxgap, unsigned int count);
+void decode_pwm(unsigned int pulses[], unsigned int gaps[], unsigned int count);
+BYTE generic_decode_pwm(BYTE *result, unsigned int pulses[], unsigned int minpulse, unsigned int maxpulse, unsigned int gaps[], unsigned int mingap, unsigned int maxgap, unsigned int count);
